FIX: parser edge cases
Fixes Array and associated Reverse Polish Notation parsing so that RPN
expressions inside of arrays are properly handled. Test case is taken from
example in OpenDSS manual, and is therefore known to be valid.

Fixes parsing of properties and their associated property names. According
to the OpenDSS manual, keyword properties reset the place in the list of
in-order properties, so that any unnamed properties following the named
one will be in the order of the list, meaning it is possible to go back to the
beginning of the property name list several times during a parse.

Adds edge cases where multiple property names are accepted for the same
property.