ADD: Support for length and units on Lines
Adds support for length and units on lines that are not `1.0` and `none`
by converting all units to meter/per meter. Removes warnings within
`src/io/opendss.jl:dss2tppm_branch!`

Adds correct support for b1,b0 in lines.

Adds stub for support for rg,xg in ines, but requires calculation of
frequency solution to work correctly, so it is not yet fully supported.

Depreciates use of `createLinecode` in parser, now only used for
parsing data types of the raw OpenDSS linecode entries, due to the
added support for Rg, Xg, which are only calculated directly in lines.

Removes test of test2_master.dss checking it solves with SOCWR,
this is only a parser test, not a real network, and only solved
incidentally before unit support was added.

Closes #43