FIX: bug in parsing ~ lines with preceeding whitespace
In opendss `~` indicates continuation of previous line, but in cases
where whitespace preceeded the `~` character, that character was
inserted into the parsed data (in the specific case where this was
discovered, into the `phases` data field of a transformer. The order
of the `strip` commands was changed to `strip(strip(line), "~")` to
solve this problem. Unit test is included.