REF: change symantics of active_phases
Renames dss_source_id struct to DSSSourceId to follow Julia naming
conventions.

Changes "active_phases" field to be a sparse array of Ints representing
each active phase, instead of the Bool convention that is similar to the
bus parser output.

Changes `active_phases` member of the struct to be a Set{Int} instead
of Array{Int}.

Changes `dss_type` and `dss_name` to be AbstractString instead of
String.