FIX: OpenDSS switch parsing
In OpenDSS, the `switch=y` property results in `length=0.001` and `units=none`.

Fixing these values results in the correct voltages across the buses, but not the correct power, which I believe must be due to a different power balance constraint being applied than lines have in OpenDSS. This issue is already covered by #156

Changelog and unit tests updated.