FIX: OpenDSS transformer %rs and %loadloss handling
Fixes the handling of `%rs` and `%loadloss` properties on OpenDSS transformers.

For 3-winding transformers:
`%loadloss = sum(%rs[1:2])`
`%rs = [%loadloss/2, %loadloss/2, 0.2]`

For 2-winding transformers, `%rs = fill(%loadloss / 2, 2)The default value `%r = 0.2` is corrected (was 0.0)

Because the of the new default values, the number of star buses created for 3-winding transformers changes.

Changelog and Unit tests updated

Closes #157