FIX: Wrong setting of basefreq on lines / linecodes (OpenDSS)
When lines / linecodes do not have the basefreq property, internally the basefreq was previous set to 60 Hz, even if the circuit basefreq is e.g. 50 Hz.

This fix changes this behavior so that if basefreq is not given on a line or linecode, the basefreq for that line is the circuit level basefreq.

The default basefreq for the circuit is still 60 Hz if `set defaultbasefreq` is not used; a warning is given if this command is missing.

Updates unit tests and changelog

Closes #159