ADD: virtual branch to sourcebus to model circuit impedance
In OpenDSS circuits connections to the sourcebus have some impedance. PowerModelsDistribution did not model this, but instead relied on cases have a very stiff source.

To resolve this, the OpenDSS parser creates a virtual sourcebus and a virtual branch with the impedance given by the circuit definition connecting to the sourcebus.

The sourcebus generator (sourcegen) has been moved to the virtual sourcebus, as it was an approximation of OpenDSS's vsource at the sourcebus.

Test Changes Notes
- For a couple cases in PF tests against pg / objective were removed, because to compare directly against OpenDSS we primarily use voltages, as that is where direct comparison can be made.
- Any voltages that were updated were taken directly from OpenDSS
- The phase-drop cases' solvability depend strongly on the stiffness of the source now that the impedance has been explicity added, so some tests had to be changed, particularly in OPF

SOCWRPowerModel from PowerModels added to exports list, as we use this during tests and it was missed previously.

Updated Changelog