FIX: phase order swap on lines
In OpenDSS, the order of phases can be swaped on lines, e.g. `bus1=a.1.2.3 bus2=b.1.2.3` vs `bus1=a.3.2.1 bus2=b.3.2.1`. This has the effect of reordering the line property matrices, i.e. `rmatrix`, `xmatrix`, and `cmatrix`.

This PR accounds for this change by obtaining the phase connections in the order that they appear and reordering the matrices accordingly.

Adds a unit test and updates changelog