FIX: apply OpenDSS properties in order written
In OpenDSS, properties are applied to their respective objects in the order they are written, which means `linecode=x, r0=1` is different than `r0=1 linecode=x`, but until now PowerModelsDistribution applied properties in a fixed order of, `like` then `linecode`, overwriting properties that shouldn't have been overwritten.

Adds additional parsing functions to make this possible. In `parse_dss`, the order properties are entered is now retained in `"prop_order"`, which is used during the conversion to the internal data format.

In OpenDSS, `like` should be applied at the beginning of a line, but its application doesn't seem consistent with the manual, so if it is not
one of the first properties given, a warning is thrown by PowerModelsDistribution that PMD will treat the `like` property as having appeared before all other keywords.

Updates changelog and adds unit test.

Closes #131