FIX: raw property parsing in OpenDSS on transformers
In cases where properties are specified, e.g.,

`transformer.name.prop1=value prop2=value ...`

specifically on transformers if `wdg` is specified there is a subset of property names that need to be changed to reflect which winding they belong to.

This fixes this problem by checking for a wdg property during the parse, and correcting setting the suffix on the property names.

This also fixes a bug where property names were not correctly set according to winding in the initial parse.