FIX: Voltage at virtual_sourcebus should be fixed
Because the sourcebus is a voltage source with an internal impedance
that we model as a virtual sourcebus connected to a virtual branch
connected to the real sourcebus, the virtual sourcebus should have
a fixed voltage.

To remedy this, `vmin` and `vmax` at the virtual_sourcebus are fixed
to `vm`, i.e. `vmin==vmax==vm`.

Only three OPF test cases needed to be updated, and none of the
voltages in those test cases needed to change:

- 2-bus diagonal acp opf
- 3-bus balanced acp opf
- 3-bus unbalanced isc acp opf

Updates unit tests and changelog