FIX: opendss tests
Fixes that enable the previous opendss tests to pass

Some tests were no longer valid, and therefore removed. For the most
part, the pointers to the correct items needed to be changed, but not
the values.

Still not passing are

- "opendss parse load model errors"
- "opendss parse generic branch values verification"
- "test json parser"

Adds line_reactor, which currently parses to a virutal branch