ADD: support for filetype autodetect
Adds support for automatically extrapolating the filetype by checking if the file can be parsed with JSON first.

Note that if filetype is not specified, and the file size is large, it may be significantly slower to parse, and therefore filetype should be specified by advanced users in complex workflows.