ADD: spectrum dss objects (#331)
This adds support for spectrum dss objects, which are commonly referenced in the manual, but the manual does not define the objects themselves, and therefore they we missed. This addition is meant to prevent the automatic failure of the dss parser when coming across these objects.

Adds spectrum objects to test2_master, and adds unit tests.