REF: json import/export
Instead of exporting a custom json format for PMD data structures, we now will export via pure JSON.print, which means that on parse we have to do some extra work to correct data types and some values (e.g. Enums, and nulls).

No unit tests needed to be updated with the refactor of this feature.