ADD: Support for LineGeometry / LineSpacing
This adds support for dss LineGeometry, LineSpacing, Wiredata objects.

It uses calculations from Kersting's "Distribution System Modeling and Analysis" to convert from LineGeometry and/or LineSpacing with Wiredata objects to series impedance and shunt admittance matrices. This calculation can be found in `_calc_impedances_from_geometry`

All units get converted to meters on parse, and therefore any equations need to also be in meters to be correct.

On LineGeometry objects, there is a feature similar to Transformers, where the "cond" can be set, and any properties set afterwards are assigned to that conductor. To support this, _add_property had to be refactored slightly to support keys and properties specific to the object type.

A bug in _parse_element_with_dtype was fixed, where values where left as Substrings instead of getting converted to String

TODO

- [ ] Verify calculations are correct
- [ ] Add unit tests
- [ ] Update changelog