[ADD][4W-EXTENSION] Parsing, transformations and solution builder (#339)
* bug in _correct_branch_directions!

* solution builder

when the data type is Any for a DenseAxisArray, call the solution builder for each element to correctly specialize for each type in the vector

* optional phase projection

added a keyword to turn of phase projection

* math transformations

Added a keyword to pass transformations at the mathematical model level (analogous to transformations in the parser). This is useful for the voltage starting value methods.

* reactor series impedance

Pass through the series impedance values. This allows for correct handling of groundings implemented as reactors.

* bug edge case

Prevents errors when a line connects only to a zero terminal. This happens for groundings.

* voltage source parsing

Prevents copying of sourcebus terminals which are not part of the voltage source (occurs because a short-circuited terminal is sometimes linked to the neutral to prevent OpenDSS from omitting it from the results).

* line reduction transformations

Adds transformations which reduce the lines in an engineering data model.

* no-load initialization

Calculates and adds no-load voltage starting values to the mathematical data model. This is key to speed up and achieve convergence in complex cases.

* line loops

Resolves line loops (lines connected to a single bus, typically result of grounding reactors which are parsed to lines).

* Kron-reduction

Applies a Kron-reduction at the data engineering model level, by detecting the implicit neutrals in the network.

* typo

* vstart refactor bug

* typo

* voltage bound transformations

* combine vbounds

data model transformation combines various  ways of defining voltage bounds to the tightest unique set covering all bounds

* bug

* pp_pn_ng vbounds

* bug fixes

* tap prop removed

* phase_projected docstring

* transformations removed

* connections edge cases

* tripple double quotes

* verbose removed

* added docstrings